+++ Source/AliveLibAE/Sound/Sound.hpp
@@ -103,7 +103,7 @@ EXPORT int CC SND_Clear_4EF350(SoundEntry* pSoundEntry, unsigned int sampleOffse
 EXPORT void CC SND_SsQuit_4EFD50();
 EXPORT signed int CC SND_CreateDS_4EEAA0(unsigned int sampleRate, int bitsPerSample, int isStereo);
 EXPORT void CC SND_Init_WaveFormatEx_4EEA00(WAVEFORMATEX *pWaveFormat, int sampleRate, unsigned __int8 bitsPerSample, int isStereo);
-EXPORT signed int CC SND_New_4EEFF0(SoundEntry *pSnd, int sampleLength, int sampleRate, int bitsPerSample, int isStereo);
+EXPORT signed int CC SND_New_4EEFF0(SoundEntry *pSnd, int sampleLength, int sampleRate, BYTE bitsPerSample, int isStereo);
 EXPORT int CC SND_Load_4EF680(SoundEntry* pSnd, const void* pWaveData, int waveDataLen);
 EXPORT const char* CC SND_HR_Err_To_String_4EEC70(HRESULT hr);
 EXPORT signed int CC SND_Free_4EFA30(SoundEntry* pSnd);
